requirements.txt
asgiref==3.7.2
Django==4.2.6
sqlparse==0.4.4
typing_extensions==4.8.0

requests~=2.31.0
PyJWT~=2.8.0
cent~=4.1.0